backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
python-multipart
sqlalchemy==2.0.46
pydantic==2.12.5
PyPDF2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic_core==2.41.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0


# ﻿annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.12.1
# click==8.3.1
# colorama==0.4.6
# dnspython==2.8.0
# email-validator==2.3.0
# fastapi==0.129.0
# greenlet==3.3.1
# h11==0.16.0
# idna==3.11
# psycopg2-binary==2.9.11
# pydantic==2.12.5
# pydantic_core==2.41.5
# SQLAlchemy==2.0.46
# starlette==0.52.1
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# uvicorn==0.41.0